chore: restructure eth module hierarchy
- moved primitive/foreign tokenizable impls into own module
- move common api type to `common` module
- moved other tokenizable impls next to struct defs
- deleted some unused types and code (impl TryInto for AggKey)
- move api stuff into api module
- generally tidied up imports
